update libraries for new Java and spring update in AMS
