+++ src/main/java/io/antmedia/enterprise/streamapp/AMSEndpointConfigurator.java
@@ -1,8 +1,8 @@
 package io.antmedia.enterprise.streamapp;
 
-import javax.websocket.HandshakeResponse;
-import javax.websocket.server.HandshakeRequest;
-import javax.websocket.server.ServerEndpointConfig;
+import jakarta.websocket.HandshakeResponse;
+import jakarta.websocket.server.HandshakeRequest;
+import jakarta.websocket.server.ServerEndpointConfig;
 
 import org.apache.tomcat.websocket.server.DefaultServerEndpointConfigurator;
 import org.slf4j.Logger;
