+++ src/main/java/io/antmedia/enterprise/streamapp/WebSocketSignalingLocalHandler.java
@@ -2,13 +2,13 @@ package io.antmedia.enterprise.streamapp;
 
 import java.io.IOException;
 
-import javax.websocket.EndpointConfig;
-import javax.websocket.OnClose;
-import javax.websocket.OnError;
-import javax.websocket.OnMessage;
-import javax.websocket.OnOpen;
-import javax.websocket.Session;
-import javax.websocket.server.ServerEndpoint;
+import jakarta.websocket.EndpointConfig;
+import jakarta.websocket.OnClose;
+import jakarta.websocket.OnError;
+import jakarta.websocket.OnMessage;
+import jakarta.websocket.OnOpen;
+import jakarta.websocket.Session;
+import jakarta.websocket.server.ServerEndpoint;
 
 import org.apache.catalina.core.ApplicationContextFacade;
 import org.apache.commons.lang3.exception.ExceptionUtils;
