--- src/main/java/io/antmedia/saml/SamlSecurityConfig.java
@@ -1,285 +1,0 @@
-package io.antmedia.saml;
-
-import org.opensaml.saml2.metadata.provider.MetadataProvider;
-import org.opensaml.saml2.metadata.provider.MetadataProviderException;
-import org.opensaml.saml2.metadata.provider.ResourceBackedMetadataProvider;
-import org.opensaml.util.resource.HttpResource;
-import org.opensaml.util.resource.ResourceException;
-import org.opensaml.xml.parse.StaticBasicParserPool;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.io.DefaultResourceLoader;
-import org.springframework.core.io.FileSystemResource;
-import org.springframework.security.saml.*;
-import org.springframework.security.saml.context.SAMLContextProviderImpl;
-import org.springframework.security.saml.key.JKSKeyManager;
-import org.springframework.security.saml.key.KeyManager;
-import org.springframework.security.saml.log.SAMLDefaultLogger;
-import org.springframework.security.saml.metadata.CachingMetadataManager;
-import org.springframework.security.saml.metadata.ExtendedMetadata;
-import org.springframework.security.saml.metadata.ExtendedMetadataDelegate;
-import org.springframework.security.saml.processor.*;
-import org.springframework.security.saml.util.VelocityFactory;
-import org.springframework.security.saml.websso.*;
-import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.security.web.authentication.logout.LogoutHandler;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
-import org.springframework.security.web.authentication.logout.SimpleUrlLogoutSuccessHandler;
-
-import java.io.File;
-import java.util.*;
-
-@Configuration
-@PropertySource("/WEB-INF/red5-web.properties")
-public class SamlSecurityConfig {
-
-    private static final Logger logger = LoggerFactory.getLogger(SamlSecurityConfig.class);
-
-    private String samlKeystoreLocation;
-    //= "/home/karinca/softwares/ant-media-server/saml/keystore.jks";
-
-    private String samlKeystorePassword;
-//= "oktaiscool";
-
-    private String samlKeystoreAlias;
-    //= "oktasaml";
-
-    private String defaultIdp;
-    //= "http://www.okta.com/exk1mfkni2bpGXA6k5d7";
-
-    private String metadataUrl;
-    //= "https://dev-75335055.okta.com/app/exk1mfkni2bpGXA6k5d7/sso/saml/metadata";
-
-
-    public String getSamlKeystoreLocation() {
-        return samlKeystoreLocation;
-    }
-
-    public void setSamlKeystoreLocation(String samlKeystoreLocation) {
-        this.samlKeystoreLocation = samlKeystoreLocation;
-    }
-
-    public String getSamlKeystorePassword() {
-        return samlKeystorePassword;
-    }
-
-    public void setSamlKeystorePassword(String samlKeystorePassword) {
-        this.samlKeystorePassword = samlKeystorePassword;
-    }
-
-    public String getSamlKeystoreAlias() {
-        return samlKeystoreAlias;
-    }
-
-    public void setSamlKeystoreAlias(String samlKeystoreAlias) {
-        this.samlKeystoreAlias = samlKeystoreAlias;
-    }
-
-    public String getDefaultIdp() {
-        return defaultIdp;
-    }
-
-    public void setDefaultIdp(String defaultIdp) {
-        this.defaultIdp = defaultIdp;
-    }
-
-    public String getMetadataUrl() {
-        return metadataUrl;
-    }
-
-    public void setMetadataUrl(String metadataUrl) {
-        this.metadataUrl = metadataUrl;
-    }
-
-    @Bean(initMethod = "initialize")
-    public StaticBasicParserPool parserPool() {
-        return new StaticBasicParserPool();
-    }
-
-    @Bean
-    public SAMLAuthenticationProvider samlAuthenticationProvider() {
-        return new CustomSAMLAuthenticationProvider();
-    }
-
-    @Bean
-    public SAMLContextProviderImpl contextProvider() {
-        return new SAMLContextProviderImpl();
-    }
-
-    @Bean
-    public static SAMLBootstrap samlBootstrap() {
-        return new SAMLBootstrap();
-    }
-
-    @Bean
-    public SAMLDefaultLogger samlLogger() {
-        return new SAMLDefaultLogger();
-    }
-
-    @Bean
-    public WebSSOProfileConsumer webSSOprofileConsumer() {
-        return new WebSSOProfileConsumerImpl();
-    }
-
-    @Bean
-    @Qualifier("hokWebSSOprofileConsumer")
-    public WebSSOProfileConsumerHoKImpl hokWebSSOProfileConsumer() {
-        return new WebSSOProfileConsumerHoKImpl();
-    }
-
-    @Bean
-    public WebSSOProfile webSSOprofile() {
-        return new WebSSOProfileImpl();
-    }
-
-    @Bean
-    public WebSSOProfileConsumerHoKImpl hokWebSSOProfile() {
-        return new WebSSOProfileConsumerHoKImpl();
-    }
-
-    @Bean
-    public WebSSOProfileECPImpl ecpProfile() {
-        return new WebSSOProfileECPImpl();
-    }
-
-    @Bean
-    public SingleLogoutProfile logoutProfile() {
-        return new SingleLogoutProfileImpl();
-    }
-
-    @Bean
-    public KeyManager keyManager() {
-        DefaultResourceLoader loader = new DefaultResourceLoader();
-        //Resource storeFile = loader.getResource(samlKeystoreLocation);
-        FileSystemResource storeFile = new FileSystemResource(new File(samlKeystoreLocation));
-        Map<String, String> passwords = new HashMap<>();
-        passwords.put(samlKeystoreAlias, samlKeystorePassword);
-        return new JKSKeyManager(storeFile, samlKeystorePassword, passwords, samlKeystoreAlias);
-    }
-
-    @Bean
-    public WebSSOProfileOptions defaultWebSSOProfileOptions() {
-        WebSSOProfileOptions webSSOProfileOptions = new WebSSOProfileOptions();
-        webSSOProfileOptions.setIncludeScoping(false);
-        return webSSOProfileOptions;
-    }
-
-    @Bean
-    public SAMLEntryPoint samlEntryPoint() {
-        SAMLEntryPoint samlEntryPoint = new SAMLEntryPoint();
-        samlEntryPoint.setDefaultProfileOptions(defaultWebSSOProfileOptions());
-        return samlEntryPoint;
-    }
-
-    @Bean
-    public ExtendedMetadata extendedMetadata() {
-        ExtendedMetadata extendedMetadata = new ExtendedMetadata();
-        extendedMetadata.setIdpDiscoveryEnabled(false);
-        extendedMetadata.setSignMetadata(false);
-        return extendedMetadata;
-    }
-
-    @Bean
-    @Qualifier("okta")
-    public ExtendedMetadataDelegate oktaExtendedMetadataProvider() throws MetadataProviderException {
-
-        // Use the Spring Security SAML resource mechanism to load
-        // metadata from the Java classpath.  This works from Spring Boot
-        // self contained JAR file.
-        org.opensaml.util.resource.Resource resource = null;
-
-
-        /*
-		try {
-            resource = new ClasspathResource("/saml/metadata/sso.xml");
-		} catch (ResourceException e) {
-			 e.printStackTrace();
-		}
-		*/
-        resource = new HttpResource(metadataUrl);
-
-        Timer timer = new Timer("saml-metadata");
-        ResourceBackedMetadataProvider provider = new ResourceBackedMetadataProvider(timer,resource);
-        provider.setParserPool(parserPool());
-        return new ExtendedMetadataDelegate(provider, extendedMetadata());
-    }
-
-    @Bean
-    @Qualifier("metadata")
-    public CachingMetadataManager metadata() throws MetadataProviderException, ResourceException {
-        List<MetadataProvider> providers = new ArrayList<>();
-        providers.add(oktaExtendedMetadataProvider());
-        CachingMetadataManager metadataManager = new CachingMetadataManager(providers);
-        metadataManager.setDefaultIDP(defaultIdp);
-        return metadataManager;
-    }
-
-    @Bean
-    @Qualifier("saml")
-    public SavedRequestAwareAuthenticationSuccessHandler successRedirectHandler() {
-        SavedRequestAwareAuthenticationSuccessHandler successRedirectHandler = new SavedRequestAwareAuthenticationSuccessHandler();
-        successRedirectHandler.setDefaultTargetUrl("/home");
-        return successRedirectHandler;
-    }
-
-    @Bean
-    @Qualifier("saml")
-    public SimpleUrlAuthenticationFailureHandler authenticationFailureHandler() {
-        SimpleUrlAuthenticationFailureHandler failureHandler = new SimpleUrlAuthenticationFailureHandler();
-        failureHandler.setUseForward(true);
-        failureHandler.setDefaultFailureUrl("/error");
-        return failureHandler;
-    }
-
-    @Bean
-    public SimpleUrlLogoutSuccessHandler successLogoutHandler() {
-        SimpleUrlLogoutSuccessHandler successLogoutHandler = new SimpleUrlLogoutSuccessHandler();
-        successLogoutHandler.setDefaultTargetUrl("/");
-        return successLogoutHandler;
-    }
-
-    @Bean
-    public SecurityContextLogoutHandler logoutHandler() {
-        SecurityContextLogoutHandler logoutHandler = new SecurityContextLogoutHandler();
-        logoutHandler.setInvalidateHttpSession(true);
-        logoutHandler.setClearAuthentication(true);
-        return logoutHandler;
-    }
-
-    @Bean
-    public SAMLLogoutProcessingFilter samlLogoutProcessingFilter() {
-        return new SAMLLogoutProcessingFilter(successLogoutHandler(), logoutHandler());
-    }
-
-    @Bean
-    public SAMLLogoutFilter samlLogoutFilter() {
-        return new SAMLLogoutFilter(successLogoutHandler(),
-                new LogoutHandler[] { logoutHandler() },
-                new LogoutHandler[] { logoutHandler() });
-    }
-
-    @Bean
-    public HTTPPostBinding httpPostBinding() {
-        return new HTTPPostBinding(parserPool(), VelocityFactory.getEngine());
-    }
-
-    @Bean
-    public HTTPRedirectDeflateBinding httpRedirectDeflateBinding() {
-        return new HTTPRedirectDeflateBinding(parserPool());
-    }
-
-    @Bean
-    public SAMLProcessorImpl processor() {
-        ArrayList<SAMLBinding> bindings = new ArrayList<>();
-        bindings.add(httpRedirectDeflateBinding());
-        bindings.add(httpPostBinding());
-        return new SAMLProcessorImpl(bindings);
-    }
-}
-
