+++ src/main/java/io/antmedia/saml/SecurityConfiguration.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2002-2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.antmedia.saml;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.saml2.provider.service.registration.InMemoryRelyingPartyRegistrationRepository;
+import org.springframework.security.saml2.provider.service.registration.RelyingPartyRegistration;
+import org.springframework.security.saml2.provider.service.registration.RelyingPartyRegistrationRepository;
+import org.springframework.security.saml2.provider.service.registration.RelyingPartyRegistrations;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfiguration {
+	private String metadataUrl;
+
+	@Bean
+	SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+		http
+				.authorizeHttpRequests((authorize) -> authorize.anyRequest().authenticated())
+				.saml2Login(Customizer.withDefaults());
+
+		return http.build();
+	}
+
+	@Bean
+	RelyingPartyRegistrationRepository relyingPartyRegistrationRepository() {
+		RelyingPartyRegistration relyingPartyRegistration = RelyingPartyRegistrations
+				.fromMetadataLocation(metadataUrl)
+				.registrationId("okta")
+				.build();
+
+		return new InMemoryRelyingPartyRegistrationRepository(relyingPartyRegistration);
+	}
+
+	public String getMetadataUrl() {
+		return metadataUrl;
+	}
+
+	public void setMetadataUrl(String metadataUrl) {
+		this.metadataUrl = metadataUrl;
+	}
+}
